fix: separate controlPanel on diff components
